build.sh: use auto detected gcc if CC=gcc is set
The motivation of this change is to be able to use the autodetected gcc
version (e.g. the bundled prebuilt) even in case CC=clang has been set
by the build.config.

It appears a bit hacky, but only breaks backward compatibility for the
case that somebody uses CC=gcc and expects the gcc from PATH to be used.
That case is still supported by passing an absolute path.

Change-Id: I537192168a27bdda9b288443de88ed45d7c4ec68
Signed-off-by: Matthias Maennich <maennich@google.com>