Update patch set 1
Patch Set 1:

(3 comments)

> Can you elaborate on why this change is needed? (I still don't fully understand the motivation).

If we want to switch the default compiler to clang for various kernels (like hikey or common kernel), we still need a way of allowing to build with the gcc toolchain. At least for some time. If we define CC=clang in the config, implementing this hack is one way of still allowing CC=gcc to use the prebuilt toolchain. As of now CC=gcc means in most cases /usr/bin/gcc.

Patch-set: 1
Reviewer: Gerrit User 1402878 <1402878@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0