Update patch set 1
Patch Set 1:

tl;dr What does this include guard help with? If nothing, I think we should remove it.

Whether the user invokes this script manually or not doesn't seem to be the issue here.

If the script is invoked by being sourced by another, the export of this guard will have no effect anyway, because it won't persist in the next invocation.

In the case the user *does* source this file manually, presumably to test changes to the file, it is IMO broken to skip parsing those changes with an include guard.

The AOSP platform envsetup.sh does not check if it has been included before, it simply repeats the inclusion process. Sometimes this may result in redundant modifications of PATH, but this doesn't seem to cause any real-world problems. I can't see a reason the kernel/build envsetup.sh would be any different.

Patch-set: 1
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>