Always export check_defconfig()
envsetup.sh is a configuration file that updates the environment variables by
sourcing the BUILD_CONFIG file, updating PATH, defining check_defconfig() etc.
The exported environment variables can be later used to build the kernel.
When a child process is executed, it inherits the exported variables, but not
the functions. Exporting functions in Bash can be done with "export -f", that
is unfortunately incompatible with other shells like Zsh.
This, combined with the evaluation of ENVSETUP_SH_INCLUDED that causes
envsetup.sh to be imported only once, generates an unwanted behavior:
- the user sources envsetup.sh
  - ENVSETUP_SH_INCLUDED is set to 1
  - check_defconfig() is not exported
- the user runs build.sh
  - build.sh tries to source envsetup.sh, which immediately returns because of
    ENVSETUP_SH_INCLUDED
  - build.sh tries to use check_defconfig(), that does not exist.

This patch solves the mentioned issue by causing envsetup.sh to return if
(ENVSETUP_SH_INCLUDED == 1) only after defining check_defconfig().  The
definition of check_defconfig() does not modify the environment variables, so
the single inclusion behavior (that I suspect is done to avoid environment
variables overwrites) is preserved.

Test: manual inspection
Bug: 138854687
Change-Id: If52b0546809d77304ed8d11c9da062e13a3c7ec5
Signed-off-by: Alessio Balsini <balsini@google.com>