Update patch set 1
Patch Set 1:

Matthias seemed protective of the include guard when we asked about it on another review. Also, the description here is a bit unclear. In what way is the compatibility improved by "always exporting" the macro? Shouldn't it be always exported because this script is always parsed, and the user shouldn't have manually sourced it?

Patch-set: 1
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>