build_abi: add support for symbol whitelisting
This adds support for specifying a symbol whitelist (also known as
Kernel Module Interface (KMI) whitelist) as build.config parameter. The
option KMI_WHITELIST needs to refer to a whitelist file relative to
$KERNEL_DIR and follows a very simple format [1]. Usually it will be a
text file with content like
  [abi_whitelist]
    symbol1
    symbol2
    ...

build_abi.sh will propagate this setting to diff_abi and dump_abi, such
that creation and comparison are parameterized correctly. Symbols not
mentioned in the whitelist will not be extracted into the dump created
by dump_abi and will subsequently not be compared during diff_abi.

There is a known issue in libabigail [2] in regards to comparison. There
are cases that diff_abi (i.e. abidiff) does not properly respect the
whitelist when comparing. In particular this gets hit if the baseline
and the new dump are created with different whitelists.

[1] https://sourceware.org/libabigail/manual/kmidiff.html#environment
[2] https://sourceware.org/bugzilla/show_bug.cgi?id=25095

Bug: 140596566
Change-Id: Id764b75b095ce5a0d09e04763c5b5ef10173aa75
Signed-off-by: Matthias Maennich <maennich@google.com>