Update patch set 14
Patch Set 14:

I haven't taken a close look yet, but if you only have to support ramdisk as root devices, does it let you simplify the set of changes in this patch? If you do, it might be easier to get +2 to merge this.

Patch-set: 14
Reviewer: Gerrit User 1446383 <1446383@85c56323-6fa9-3386-8a01-6480fb634889>