Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> You should already have the build number which maps back to the manifest branch.

The referenced bug claims that this information can't be derived from the kernel build. And I can also not see how this goes into the binary. (We can add the build number in addition if available though.)

It seems the only way that is available currently is to use `git log` on the binary and hope the commit message contains a valid reference to the branch/build or whatever is needed to get to the sources.

I am agreeing that this is often not consistent and hard to track down (have been there). Hence I suggest this generic way to embed this information consistently into every Android kernel binary.

Patch-set: 1
Reviewer: Gerrit User 1402878 <1402878@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0