Update patch set 1
Patch Set 1:

(41 comments)

As promised, here is the second half of the review. As written in one file comment, I think we need to find a way to make this a bit more easier to review. I know that regexes etc. have the tendency to be hard to reason about.

I am also concerned how we will integrate this into the the existing (Android) Kernel Build and the existing ABI tooling. Can you please share your ideas about this with me?

Lastly, we need to reduce (or encapsulate) the dependencies to the Kernel build internals. These might change without notice from version to version and we might see subtle bugs because of this.

Patch-set: 1
Reviewer: Gerrit User 1402878 <1402878@85c56323-6fa9-3386-8a01-6480fb634889>