add kmi_defines.py program
This is the first half of the reimplementation of the LKA tools into
a single Python 3 script:
    kmi_defins.py
For now the LKA tools are left here for reference, no need to review
the lka-* files as those will all go away.

Factored the use of subprocess.run() into its own function in preparation
for the code that comes in the patchset after this one.  Also updated checks
and comments about the form of the cc_list and the indexes into it.

Bug: 139885116
Test: manually, all code passes pylint3, ptype, and formatted with yapf

Signed-off-by: Ramon Pantin <pantin@google.com>

Change-Id: I0cdda58aa8471b8894a0d9272f8c08d3b6cd6836