add kmi_defines.py program
This is the first half of the reimplementation of the LKA tools into
a single Python 3 script:
    kmi_defins.py
For now the LKA tools are left here for reference, no need to review
the lka-* files as those will all go away.

A chunk_size of 0 (miscalculated) causes the Pool() work to return a
list of Nones.  This is a bug fix to patch set 31

Bug: 139885116
Test: manually, all code passes pylint3, ptype, and formatted with yapf

Signed-off-by: Ramon Pantin <pantin@google.com>

Change-Id: I0cdda58aa8471b8894a0d9272f8c08d3b6cd6836