add kmi_defines.py program
Program to examine a Linux build and produce the set of headers relevant
to ABI and the compile time #define constants contained in those headers
so that they can be tracked and ensure that changes to them do not break
the Kernel ABI for AOSP GKI.

This is the first half of the code. The second half of the code will be
submitted incrementally to facilitate review.

Bug: 139885116
Test: manually, all code passes pylint3, ptype, and formatted with yapf

Signed-off-by: Ramon Pantin <pantin@google.com>

Change-Id: I0cdda58aa8471b8894a0d9272f8c08d3b6cd6836