Update patch set 1
Patch Set 1: Code-Review-1

abidiff returns an exit status in the form of a set of bits (as far as I can tell). The code in the exception check only tests against the values 1 and 2 but should be doing bit-testing instead.

Patch-set: 1
Reviewer: Gerrit User 1625880 <1625880@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1