Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> abidiff returns an exit status in the form of a set of bits (as far as I can tell). The code in the exception check only tests against the values 1 and 2 but should be doing bit-testing instead.

You are right, yet I believe the handling is correct. Can you confirm?

The outcomes that we need to consider are

0    -  all good
1,2  -  crash, user error (our fault)
4    -  ABI change
8    -  incompatible change
12   -  both of the above

we fall through everything for rc=0 and return the initial value for `abi_changed` (False)

for rc!=0 we hit the except block were we handle 1,2 by reraising and everything else (including 8,4,12) is handled by eventually returning True (as in abi_changed=False)

The behaviour that changes is that we actually write a short report unconditionally (empty, but existent) for the rc=0 case and shortened for rc={4,8,12} where we exited too early before.

Patch-set: 1
Reviewer: Gerrit User 1402878 <1402878@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0