abi: add flatten_whitelist helper script
The format used by the kernel for the symbol whitelist is a simple text
file containing the list. Libabigail, on the other hand, uses a richer
INI format. So, to provide such a whitelist to the kernel build system,
a conversion step is required first.

Add a simple python script helper able convert a whitelist file from the
libabigail format to raw text.

Bug: 148277666
Change-Id: I53fe074b8f64e485697b607fcd626990d8c270bb
Signed-off-by: Quentin Perret <qperret@google.com>