kmi_defines: Revamped command line option processing
This addresses support for running the program without multiprocessing and
to dump information only when requested.

Headers relevant to ABI are now selected by rule that a header is ABI
relevant if it is used by the kernel and one kernel module or by two
kernel modules (an inter-module ABI).

Bug: 139885116
Change-Id: I03df624fac4bb87b132fb4c3b67e4e7c6624b146
Signed-off-by: Matthias Maennich <maennich@google.com>