GKI: devices_snapshot noise reduction
For the /sys difference report, if /sys/module/<name> is introduced
clean, then replace the long diff reporting each element with a
single mention of /sys/module/<name>/... so that the developer is
not confronted with a lot of spurious details.

We also deal better with failure to pull data, eg: device fails to
boot, leaving an empty file, we now delete it and filter diff spam
out.

Finally, we squash major changes in module sections: bss, data,
rodata and text.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Test: manual
Bug: 147894928
Change-Id: Ic971fb25b74742dbf07860f14c09d34e15fbc88b