kmi_defines: cleanup up Target.__init__()
Removed well known indexes into cc_list, if options are
passed redundantly it is an error.

Using shlex.split() instead of trying to deal (poorly)
with shell quoting which is what the prior code did.

Moved the bulk of the work of Target.__init__() to
a new function get_cc_list().

Bug: 139885116
Change-Id: I609deb3e1cfcdc78d2ec2a5026a946f093fcc78c