kmi_defines: cleanup up Target.__init__()
Well known option indexes into cc_list are now hints
that are verified at run time and the options found
if they are not at the hinted locations.

Using shlex.split() instead of trying to deal (poorly)
with shell quoting which is what the prior code did.

Moved the bulk of the work of Target.__init__() to
a new function get_cc_list().

Change-Id: I609deb3e1cfcdc78d2ec2a5026a946f093fcc78c