kmi_defines: initialize multiprocessing, logging, and add options
Use the multiprocessing fork server to reduce the code of
the tool, the main process grows a good amount with its
accumulated state, so its best to fork the fork server
as soon as possible.

Initialized logging to include time stamps and the program
name.  Also added option to set the log level to INFO.

Added various options used by subsequent commits.

Bug: 139885116
Change-Id: I6641b0c9bab92bb353b2dd201bee79014a4308e6