kmi_defines: added list of headers to exclude and some logging
Added a list of well known headers that interfere with how this
tool works, mostly because they expect the files that include
them to pre-define certain values.  The rationale for each header
is included in the new file:
	README-kmi_defines.md
which will later become a well form .md file. For now its contents
is what matters.

Also added some logging and changed how abi_headers is computed
to take into account the exclued headers.

Change-Id: I388f1d86c80f6530856984bd6986e910e83f2757