Update patch set 2
Patch Set 2:

(3 comments)

The reason I added this check was because I kept being bitten if I ran it from the wrong window (within having the prebuilt compilers in the PATH) and the default gLinux clang was used instead (which leads to errors with asm goto constructs in the Linux inlined code).  So I wanted to spare myself (and anybody else who might run this) that confusion.

Patch-set: 2