GKI: add gki-cherry-pick
gki-cherry-pick script gets handed a list of shas, and kernel git
trees, and evaluates which shas need to be handed to each kernel
based on files referenced and formulates a per-tree single
split.patch file that can be fed to git am -3 -k to apply.

The series is provided in split.patch in source order, so the shas
supplied need not be in any order.  Also, the shas are investigated
for title matches so that existing patches are cancelled out.  The
script does this research for you, at about 6 seconds per sha per
kernel.

Patches that can be applied as-is are presented as cherry picks.
Patches that need to be split up are presented as new content that
references the original patch and what actions were taken to make
it partial.  Of course the conflict resolutions must be managed
incrementally as git am goes through the series.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Change-Id: Ic8bbaa91c58f6fc5e670ac0aad881adeb0d2b688