kmi_defines: removed unneeded pytype comments
Removed comments to cause pytype not to complaint about a few
classes that in prior versions of the code had to few methods.

Removed pytype comment about unused import of Set, in earlier
version Set was only used for typing checks (pytype), but was
unused otherwise, now it is used for other purposes.

Change-Id: I75c8e89a71c62c1726c02bfcb99db37bc2b0683c
Signed-off-by: Ramon Pantin <pantin@google.com>
Test: interactively
Test: pytype kmi_defines.py
Test: pylint3 kmi_defines.py
Bug: 139885116