abi: add gki_check tool to easy GKI related ABI comparisons
gki_check will download the correct latest ABI representation for
android-5.4 or android-4.19 and facilitate the existing tools dump_abi
and diff_abi to have a one step comparison against a locally built
kernel. The ABI representation download is cached to not introduce
unnecessary bandwidth usage.

A sample invocation to compare a locally built kernel tree is as
follows:

 $ build/abi/gki_check --linux-tree path/to/out/ --kernel-version 5.4

Alternatively, kernel version 4.19 is supported and a KMI whitelist can
be passed as needed.

TODO: Add documentation

Bug: 149423459
Change-Id: I4cc19b953d92232d5af8aac2bb3f6510582744f2
Signed-off-by: Matthias Maennich <maennich@google.com>