Update patch set 1
Patch Set 1:

I have the usual preference to prefer positive flag names.

So I'd suggest: --module-grouping, --no-module-grouping and set the default to true. This will mean an extra "not" somewhere else.

Also, in the conditional logic, I'd swap the conjunction as the exported check is fundamental.
So: if symbol in exported and (count > 1 or not module_grouping).

These are not big issues though.

Patch-set: 1
Reviewer: Gerrit User 1625880 <1625880@85c56323-6fa9-3386-8a01-6480fb634889>