Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I have the usual preference to prefer positive flag names.
> 
> So I'd suggest: --module-grouping, --no-module-grouping and set the default to true. This will mean an extra "not" somewhere else.

The default will be the grouping enabled, so the flag needs to disable it. That is why I named it --skip-module-grouping. The logic within the implementation avoids the double negative though. I can rename it to --no-module-grouping. How do you suggest naming the flag to disable grouping with keeping it positive?

> 
> Also, in the conditional logic, I'd swap the conjunction as the exported check is fundamental.
> So: if symbol in exported and (count > 1 or not module_grouping).

I would rather not reorder this as 'symbol in exported' is a map lookup and is much more expensive compared to 'count > 1 or not module_grouping'.

> 
> These are not big issues though.

Patch-set: 1
Reviewer: Gerrit User 1402878 <1402878@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0