Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > +rprichard is looking at switching us over to the LLVM EH unwinder from libgcc. i don't know whether the host is extra work, or we'll just get that "for free" when x86-64 is switched over...
> 
> That means that you currently still use the libgcc* etc. from the host?

For Soong linux-x86_64 host builds, it looks like we invoke the clang++ driver to link, and we pass libclang_rt.builtins-x86_64.a. There is a -nostdlib++ but not -nostdlib. I don't see a libgcc or libgcc_s library mentioned explicitly, but they're added implicitly after libclang_rt.builtins.

i.e. I think we're mostly relying on compiler-rt, but the unwinder should be coming from libgcc_s.so. (I don't know if libgcc.a is used at all.)

Patch-set: 2