Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: Code-Review+1
> > > 
> > > build.sh change LGTM.  We set HOSTLD in our upstream CI.
> > > https://github.com/ClangBuiltLinux/continuous-integration/blob/master/driver.sh#L323
> > > 
> > > I'm not really sure what the ld file is you've added, or if the Docker image change is ok.  Maybe you can have someone else help review those changes?
> > 
> > The problem is that the build tries to use 'ld' event hough HOSTLD=ld.lld is set. So, it breaks without the little ld script. I suspect that this is a bug as the build should not rely on a host ld if HOSTLD=ld.lld.
> 
> That sounds like a workaround for part of the build that's not hermetic, that we should find and fix upstream?

Yes, sorry, I was not clear on my initial ask. (I should have opened a bug instead.) The container setup that I have here is an easy (and quick) way to reproduce the issue.  (./run.sh)
Dropping the 'ld' script makes `fixdep` fail to link due to missing 'ld'.

There are more solutions to this (e.g. adding ld->ld.lld as a symbolic link to the clang prebuilts solves the hermetic problem for *us*, but does not solve the actual upstream problem).

Patch-set: 2