RFC: Add a script to verify consistent ABI for UAPI header changes
This is a proposal for a script that verifies that there are no UABI
changes in a given GIT commit.

The script extracts any file in a uapi/ directory from the commit and
constructs a sanitized 'before' and 'after' version of the file.  A
simple binary is build with both versions of the file and compared with
abidiff.

This depends on GCC with -fno-eliminate-unused-debug-types to ensure that
all the DWARF objects are retained in the binary.  It also expects that
unifdef and abidiff are in the path.  The script itself is intended to
be run from the kernel base directory.

Bug: 73298553
Change-Id: Ib6f3e2191a2bf82a97f829854c1c3f2bdc3a83ca
Signed-off-by: Jordan Crouse <jcrouse@quicinc.com>