Update patch set 2
Patch Set 2:

(4 comments)

1) I had our downstream model in mind when I wrote this specific script and we run verification on each commit. I'll investigate the idea of building a big library. My guess is that we'll need to blacklist some of the common headers (linux/types.h maybe?) to make it work, but that wouldn't be a big deal.

But how would that work for an incremental build? Say you build a big library and then change qrtr.h - would you need to build the entire library for a comparison, or just qrtr.h and if you did that wouldn't abidiff complain about all the missing debug symbols?  I'll have to play with it a bit.

2) I'm a newbie to this project, so I'll need to investigate those scripts a bit more.

3) I can do it in python. I originally had it in perl because I'm old and stubborn but everything piped together so well it seemed appropriate to a shell script,and also I didn't need to translate the headers_install.sh sed script. 😊  Let me see if I can figure out how to build a comprehensive uapi library first and then go from there.

Patch-set: 2