abi/extract_symbols: Add option to generate per-module whitelists
Passing --emit-module-whitelists to abi/extract_symbols, makes it emit
individual whitelists for each processed module. It does so in addition
to emitting to the specified --whitelist and with filenames like
whitelist + "_" + module_name. The module specific whitelists
intentionally have overlaps among them as they also list the commonly
used symbols that are combined in the main whitelist.

Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I954aad4ccfa4613049d8648e8dcaacdf2f475a14