Find depmod on more distros
When running build.sh on modern distros, an error may be seen:

build/build.sh: line 471: depmod: command not found

This occurs because /sbin and /usr/sbin are not in PATH for non-root
users by default. Kbuild's use of this tool already handles this case,
but we have another open-coded use which does not.

Test: ran build/build.sh on Debian Bullseye (testing)
Change-Id: I51088f93322fd1387ae9abd953779e9018733f2a
Signed-off-by: Alistair Delva <adelva@google.com>