Update patch set 2
Patch Set 2:

I think it would be better if this wrote the abi.prop file in one go. There is a risk of a partial file being written and confusing people if the script exits midway.

Have a think about what the lifetime of the abi.prop file should be.

Should it be deleted at some point during execution?
When is a good time to write it?

Also consider using a here-document and emitting the various items unconditionally (options =0 if not off?).

Patch-set: 2
Reviewer: Gerrit User 1625880 <1625880@85c56323-6fa9-3386-8a01-6480fb634889>