Useful message about environment.
The "have you run bootstrap" message will occur if a user runs
build.sh without the right environment.

Running bootstrap will offer to rebuild elfutils and libabigail, but
if they are already built, it will warn instead and give the user a
chance to abort.

This commit helpfully adds the environment needed to the warning
message so they don't need to reconstruct it or rebuild the tools.

Change-Id: I17e39e91fe3da60bfb70610140459fdb7f56200d
Signed-off-by: Giuliano Procida <gprocida@google.com>