build_abi.sh: rerun kernel build after whitelist update
When running build_abi.sh with --update and in case the main whitelist
receives an update during this, we need to rerun the build.

The main reason is that we do not know about the mechanisms to create a
full whitelist (including the additional ones). In order to do that we
need to run build.sh again.

Additionally, the KMI whitelists have an influence on the final kernel
binary (e.g. by trimming symbols to only whitelisted ones). That is an
additional reason to rerun the whole build.

Since we started from a clean dir during the first run, the second run
can be made with SKIP_MRPROPER=1 to speed things up a bit (do not expect
miracles!).

The rerun of the script in this case was necessary anyway, the rebuild
is unavoidable. Since the rebuild is the only extra step, this is the
best we can currently do giving how we built the kernel.

Bug: 151443358
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I028cad6c915999d165df4e531669546053f35f99