GKI: device_snapshot: also collect logcat -b kernel
In kernel dmesg buffer is typically not enough on a noisy development
build, so also extract the logcat -b kernel logs.  Collect dmesg and
logd.klogd immediately in the backgroup rather than waiting until
after /sys and /dev have been pulled and analyzed.

Update the most likely usage example, and add it to the script's help.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 150154308
Change-Id: I33065775713291c7f232b0706c847f5db01d0d5f