Update patch set 3
Patch Set 3:

For context, did the change to the Android platform core build for ramdisk.img land to ensure this is also lz4 compressed? In a real system, the ramdisk from vendor_boot will come from the platform build, not the kernel build, and it will be concatenated with the boot.img ramdisk, so everything has to be the same compression format.

Patch-set: 3
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>