build.sh: Allow lz4 and gzip for VENDOR_RAMDISK_BINARY
Previously, VENDOR_RAMDISK_BINARY was expected to be compressed using
the same compression format as the output vendor ramdisk image. A
mismatch resulted in an unbootable image. For example, if
VENDOR_RAMDISK_BINARY was gzip compressed and LZ4_RAMDISK=1 was set,
build.sh lz4 compressed an already gzip compressed ramdisk image.

Let's be more flexible and allow VENDOR_RAMDISK_BINARY to be in either
compression format. Also, let's print an error message if we can't
identify the compression format rather than making the assumption that
it is identical to the output compression format.

Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: Iab22acf9be1b4edbadd3d014247f3b9e4cc71001