build.sh: Check expected modules after build
In contexts such as GKI, it may be useful to keep an up-to-date list of
modules built for a given configuration. To support such a use-case,
introduce a new configuration option to build.sh called $MODULES_LIST.

$MODULES_LIST must point to a file which will be compared with the
modules.order file generated as part of the build, and any differences
will be reported as a build error.

Bug: 158655966
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: I5275112b6c2bac288647f87b601f3da798564b4c