build.sh: Archive the GKI modules list
For GKI 2.0, we need to be able to differentiate vendor and GKI modules,
both when loading modules in the kernel, and when using the ABI
monitoring tools as the GKI module ABI will need a special treatment.
In both cases, one of the pre-requisites is to have an exhaustive list
of GKI modules that can be fed to the tools.

To maintain such a list, introduce a new IS_GKI_BUILD option to build.sh
that should be set only for vanilla GKI builds, but not in downstream
vendor builds.

When set, the list of in-kernel modules for the build will be checked
against a list stored in the kernel tree, and any difference will be
reported. This is expected to be enabled in the GKI build.config
targets and to be run by TH during presubmit in ACK.

Bug: 158655966
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: I5275112b6c2bac288647f87b601f3da798564b4c