GKI: add ability to update kernel prebuilts to download_from_ci
Extend the download_from_ci script to include functionality to
update the kernel prebuilts in the Android platform build.

Updating the kernel prebuilts is done by passing the -u flag to
download_from_ci. The ANDROID_BUILD_TOP environment variable must
be set; the GKI kernel prebuilts will be updated in this platform
build. When updating the kernel prebuilts a bug number must be
supplied with -b.

The script determines the kernel version of the provided build ID
and updates the corresponding location in the Android platform,
such as kernel/prebuilts/5.4/arm64, and commits the changes in that
project. A bug tag is added using the supplied bug number.

Change-Id: I5848b4e9760ecb013eea317b43b76a618b787528