_setup_env.sh: set default value for KERNEL_DIR
to the directory where the BUILD_CONFIG is stored,
when KERNEL_DIR is not specified in the build.config*
explicitly.

By this way, we could to reduce settings of KERNEL_DIR
in build.config* when we use the build.config* in the kernel,
and the kernel could be placed in any path instead of fixed path(like common).

But for cases that the BUILD_CONFIG and the kernel source
are not in the same directory, KERNEL_DIR should be specified
explicitly in the build config file.

Test: not tested yet.

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: I75a4d82684d9b7c908583bd5da3a7429fcf6257e