_setup_env.sh: set default value for KERNEL_DIR
to the directory where the BUILD_CONFIG is stored,
when KERNEL_DIR is not specified in the build.config*
explicitly.

By this way, we could to reduce settings of KERNEL_DIR
in build.config* when we use the build.config* in the kernel,
and the kernel could be placed in any path instead of fixed path(like common).

But for cases that the BUILD_CONFIG and the kernel source
are not in the same directory, KERNEL_DIR should be specified
explicitly in the build config file.
Like for the cuttlefish-modules/build.config.cuttlefish.aarch64, need to update like this:
    $ git diff
    diff --git a/build.config.cuttlefish.aarch64 b/build.config.cuttlefish.aarch64
    index 7ac4d16..44e86a4 100644
    --- a/build.config.cuttlefish.aarch64
    +++ b/build.config.cuttlefish.aarch64
    @@ -1,6 +1,7 @@
    -. ${ROOT_DIR}/common/build.config.common
    -. ${ROOT_DIR}/common/build.config.aarch64
    -. ${ROOT_DIR}/common/build.config.gki
    +KERNEL_DIR=common
    +. ${ROOT_DIR}/${KERNEL_DIR}/build.config.common
    +. ${ROOT_DIR}/${KERNEL_DIR}/build.config.aarch64
    +. ${ROOT_DIR}/${KERNEL_DIR}/build.config.gki

     BUILD_INITRAMFS=1
     EXT_MODULES="cuttlefish-modules"
    $

Test: not tested yet.

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: I75a4d82684d9b7c908583bd5da3a7429fcf6257e