RFC: Add a script to verify consistent ABI for UAPI header changes
Add a script to generate a ABI XML file for the complete corpus of UAPI
headers (minus a few misbehaving ones).  This can be used to verify that
changes to the UAPI headers don't break backwards compatibility.

This feature depends on a compiler that supports
-fno-eliminate-unused-debug-types to make sure that all the dWARF objects
are retained from the headers.

Bug: 73298553
Change-Id: Ib6f3e2191a2bf82a97f829854c1c3f2bdc3a83ca
Signed-off-by: Jordan Crouse <jcrouse@quicinc.com>